+++ Boids/GameScene.swift
@@ -10,45 +10,18 @@ import GameplayKit
 
 class GameScene: SKScene, SKPhysicsContactDelegate {
     
-    //Settings
-    let nodeCount = 1500
-    let minimalDetectionRange: CGFloat = 30
-    let separationDistance: CGFloat = 40
-    
-    //Tree
-    let treeSubdivisionTreshold: Int = 20 // 20 for 1500 nodes
-    let treeUpdateCount = 5//Tree updates every N frames
-    
-    
-    //Modifiers
-    let cohesionModifier: CGFloat = 0.1
-    let separationModifier: CGFloat = 0.5
-    let alignmentModifier: CGFloat = 0.2
-    
-    //World settings
-
-    let maxSpeed: CGFloat = 600
-    let minSpeed: CGFloat = 500
-    let maxForce: CGFloat = 100
-    let randomness: CGFloat = 25
-    let passiveAcceleration: CGFloat = 1.1
-    
-    let borderMargin: CGFloat = 25
-    let nodeSide: CGFloat = 25
     
     //Global variables
     var borderFrame: CGRect = .zero
     var nodeArray: [BoidNode] = []
+    var updateCount: Int = 0
+    
     var previousNodeTree: QuadTree!
     var currentNodeTree: QuadTree!
-    var updateCount: Int = 0
+    
+    let nodeTexture = SKTexture(image: UIImage(named: "triangle")!) //Using the same texture to reduce amount of draws. See https://developer.apple.com/documentation/spritekit/nodes_for_scene_building/maximizing_node_drawing_performance
 
     
-    //Debug
-    let debugMode = false
-    let pauseOnStart = false
-    let zoomOut = false
-
     //MARK: - DidMove
     override func didMove(to view: SKView) {
         
@@ -58,6 +31,9 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
             .applying(scalingTransformation)
             .offsetBy(dx: nodeSide*4, dy: nodeSide*4)
             .insetBy(dx: -nodeSide*8, dy: -nodeSide*8)
+        
+        previousNodeTree = QuadTree(bounds: frame, subdivideTreshold: treeSubdivisionTreshold, minSubdivisionLenght: minimalDetectionRange)
+        currentNodeTree = QuadTree(bounds: frame, subdivideTreshold: treeSubdivisionTreshold, minSubdivisionLenght: minimalDetectionRange)
         
         //DEBUG-PART
         if zoomOut {
@@ -69,14 +45,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
         }
         //DEBUG-PART
 
-        
-        previousNodeTree = QuadTree(bounds: frame, subdivideTreshold: treeSubdivisionTreshold, minSubdivisionLenght: minimalDetectionRange)
-        currentNodeTree = QuadTree(bounds: frame, subdivideTreshold: treeSubdivisionTreshold, minSubdivisionLenght: minimalDetectionRange)
-        
-        for _ in 0..<nodeCount {
-            addNode()
-        }
-        
+        //Physics of the world
         physicsWorld.contactDelegate = self
         
         physicsBody = SKPhysicsBody(edgeLoopFrom: borderFrame)
@@ -85,6 +54,10 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
         physicsBody?.collisionBitMask = 2
         physicsBody?.categoryBitMask = 0
         
+        
+        for _ in 0..<nodeCount {
+            addNode()
+        }
         
         //DEBUG-PART
         if pauseOnStart {
@@ -99,11 +72,11 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
     
     //MARK: - Node creation
     func addNode() {
-        let node = BoidNode(size: CGSize(width: nodeSide, height: nodeSide))
+        let node = BoidNode(size: CGSize(width: nodeSide, height: nodeSide), texture: nodeTexture)
         
         //Location
-        let randomX = CGFloat.random(in: borderFrame.minX + nodeSide * 2...borderFrame.maxX - nodeSide * 2)
-        let randomY = CGFloat.random(in: borderFrame.minY + nodeSide * 2...borderFrame.maxY - nodeSide * 2)
+        let randomX = CGFloat.random(in: frame.minX + nodeSide * 2...frame.maxX - nodeSide * 2)
+        let randomY = CGFloat.random(in: frame.minY + nodeSide * 2...frame.maxY - nodeSide * 2)
         
         let randomPosition = CGPoint(x: randomX, y: randomY)
         node.position = randomPosition
@@ -118,18 +91,6 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
         nodeArray.append(node)
         previousNodeTree.addNode(node: node)
         addChild(node)
-        
-        //Giving node correct preferences
-        node.maxForce = maxForce
-        node.maxSpeed = maxSpeed
-        node.minSpeed = minSpeed
-        node.minimalDetectionRange = minimalDetectionRange
-        node.alignmentModifier = alignmentModifier
-        node.cohesionModifier = cohesionModifier
-        node.separationModifier = separationModifier
-        node.separationDistance = separationDistance
-        node.passiveAcceleration = passiveAcceleration
-        node.randomness = randomness
     }
     
     //MARK: - Contact
@@ -178,8 +139,6 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
    
     }
     
-    
-    
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         for t in touches { self.touchDown(atPoint: t.location(in: self)) }
     }
@@ -200,189 +159,30 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
     
 
     override func update(_ currentTime: TimeInterval) {
-        
         updateCount += 1
         
-        for node in nodeArray {
-//            nodeTree.remove(node)
-            let searchRect = node.getSearchRect()
-            let neighbours = self.previousNodeTree.search(searchRect: searchRect)
-            
-            node.updateValues(neighbours: neighbours) //All rules logic is inside of BoidNode class
-        }
-        
-        if updateCount == treeUpdateCount {
+        if updateCount % numberOfFramesBeforeUpdatingInfo == 0 {
             updateCount = 0
+            
             currentNodeTree.clear()
 
             for node in nodeArray {
                 self.currentNodeTree.addNode(node: node)
+
+                let searchRect = node.getSearchRect()
+                let neighbours = self.previousNodeTree.search(searchRect: searchRect)
+                node.setNeighbours(neighbours: neighbours) //All rules logic is inside of BoidNode class
+                node.updateValues()
             }
-            
             previousNodeTree = currentNodeTree
-        }
-    }
-}
 
-//MARK: - Extensions
-extension CGFloat {
-    func rad() -> CGFloat {
-        return self * (.pi / 180)
-    }
-    
-    func grad() -> CGFloat {
-        return self * (180 / .pi)
-    }
-}
-
-extension SKNode {
-    func rotateForce(by degrees: CGFloat) {
-        let oldVelocity = physicsBody!.velocity
-        
-        let myCos = cos(degrees)
-        let mySin = sin(degrees)
-
-        let rotationMatrixColumn1 = simd_double2(Double(myCos), -Double(mySin))
-        let rotationMatrixColumn2 = simd_double2(Double(mySin), Double(myCos))
-        let rotationMatrix = simd_double2x2(rows: [rotationMatrixColumn1,
-                                                      rotationMatrixColumn2])
-
-        let vectorMatrix = simd_double2(Double(oldVelocity.dx), Double(oldVelocity.dy))
-
-        let rotatedVector = matrix_multiply(rotationMatrix, vectorMatrix)
-
-        
-        let newVelocity = CGVector(dx: rotatedVector.x, dy: rotatedVector.y)
-        physicsBody?.velocity = newVelocity
-    }
-    
-    func move(toPoint: CGPoint) {
-        let move = SKAction.run {
-            self.position = toPoint
-        }
-        run(move)
-    }
-}
-
-extension SKScene {
-    func nodesInRange(point: CGPoint, range: CGFloat) -> [SKNode] {
-        return self
-            .children
-            .filter {
-                let distanceToNode = $0.position.distance(point: point)
-                return distanceToNode <= range && distanceToNode != 0
+        } else {
+            for node in nodeArray {
+                let searchRect = node.getSearchRect()
+                let neighbours = self.previousNodeTree.search(searchRect: searchRect)
+                
+                node.updateValues()
             }
-    }
-}
-
-extension CGPoint {
-    func distance(point: CGPoint) -> CGFloat {
-        return CGFloat(hypotf(Float(point.x - self.x), Float(point.y - self.y)))
-    }
-    
-    static func /=(point: inout CGPoint, value: CGFloat) {
-        point.x /= value
-        point.y /= value
-    }
-    
-    static func +=(lpoint: inout CGPoint, rpoint: CGPoint) {
-        lpoint.x += rpoint.x
-        lpoint.y += rpoint.y
-    }
-}
-
-extension CGVector {
-    static func /=(vector: inout CGVector, value: CGFloat) {
-        vector.dx /= value
-        vector.dy /= value
-    }
-    
-    static func /(vector: CGVector, value: CGFloat) -> CGVector {
-        return CGVector(dx: vector.dx / value,
-                        dy: vector.dy / value)
-    }
-    
-    static func +=(lvector: inout CGVector, rvector: CGVector) {
-        lvector.dx += rvector.dx
-        lvector.dy += rvector.dy
-    }
-    static func *=(lvector: inout CGVector, rvector: CGVector) {
-        lvector.dx *= rvector.dx
-        lvector.dy *= rvector.dy
-    }
-    static func *=(lvector: inout CGVector, rvalue: CGFloat) {
-        lvector.dx *= rvalue
-        lvector.dy *= rvalue
-    }
-    
-    static func +(lvector: CGVector, rvector: CGVector) -> CGVector {
-        return CGVector(dx: lvector.dx + rvector.dx,
-                        dy: lvector.dy + rvector.dy)
-        
-    }
-    static func -(lvector: CGVector, rvector: CGVector) -> CGVector {
-        return CGVector(dx: lvector.dx - rvector.dx,
-                        dy: lvector.dy - rvector.dy)
-        
-    }
-    static func *(lvector: CGVector, rvector: CGVector) -> CGVector {
-        return CGVector(dx: lvector.dx * rvector.dx,
-                        dy: lvector.dy * rvector.dy)
-        
-    }
-    
-    static func *(lvector: CGVector, rvalue: CGFloat) -> CGVector {
-        return CGVector(dx: lvector.dx * rvalue,
-                        dy: lvector.dy * rvalue)
-        
-    }
-    
-    static func -=(lvector: inout CGVector, rvector: CGVector) {
-        lvector.dx -= rvector.dx
-        lvector.dy -= rvector.dy
-    }
-    
-    static func -=(lvector: inout CGVector, rvalue: CGFloat) {
-        lvector.dx -= rvalue
-        lvector.dy -= rvalue
-    }
-    
-    init(randomIn: Range<CGFloat>) {
-        let randomDirection = CGVector(dx: Double.random(in: -100.0...100.0),
-                                    dy: Double.random(in: -100.0...100.0))
-        let randomForce = CGFloat.random(in: randomIn)
-        self = randomDirection.normalized() * randomForce
-    }
-
-    
-    func angleRadians() -> CGFloat {
-        return atan2(dy, dx)
-    }
-    
-    func angleDegrees() -> CGFloat {
-        return angleRadians() * 180.0 / .pi
-    }
-    
-    var lenght: CGFloat {
-        return sqrt(dx*dx + dy*dy)
-    }
-    
-    func normalized() -> CGVector {
-        let calculatedLenght = lenght
-        return calculatedLenght > 0 ? self / calculatedLenght : .zero
-    }
-    
-    mutating func limit(_ forceRange: ClosedRange<CGFloat>) {
-        if lenght > forceRange.upperBound {
-            self = normalized() * forceRange.upperBound
-        } else if lenght < forceRange.lowerBound {
-            self = normalized() * forceRange.lowerBound
-        }
-    }
-    
-    mutating func limit(_ force: CGFloat) {
-        if lenght > force {
-            self = normalized() * force
         }
     }
 }
